feat: config i18n and locales
