+++ src/app/[locale]/layout.tsx
@@ -1,0 +1,44 @@
+import '@mantine/core/styles.css';
+import { MantineProvider, ColorSchemeScript } from '@mantine/core';
+import { NextIntlClientProvider } from 'next-intl';
+import { getMessages } from 'next-intl/server';
+import { theme } from '../../../theme';
+import SplashScreenProvider from '@/src/providers/SplashScreenProvider';
+
+import '../../../styles/global.css';
+
+export const metadata = {
+  title: 'Vanessa Giannoni aka Vans',
+  description: 'Vanessa Giannoni aka Vans portfolio',
+};
+
+interface RootLayoutParams {
+  children: React.ReactNode;
+  params: {
+    locale: string;
+  };
+}
+
+export default async function RootLayout({ children, params }: RootLayoutParams) {
+  const messages = await getMessages();
+
+  return (
+    <html lang={params.locale}>
+      <head>
+        <ColorSchemeScript />
+        <link rel="shortcut icon" href="/vans-logo.png" type="image/x-icon" />
+        <meta
+          name="viewport"
+          content="minimum-scale=1, initial-scale=1, width=device-width, user-scalable=no"
+        />
+      </head>
+      <body>
+        <NextIntlClientProvider messages={messages}>
+          <MantineProvider theme={theme}>
+            <SplashScreenProvider>{children}</SplashScreenProvider>
+          </MantineProvider>
+        </NextIntlClientProvider>
+      </body>
+    </html>
+  );
+}
