+++ src/constants/locales.ts
@@ -1,0 +1,1 @@
+export const locales = ['en', 'es', 'pt'];
