+++ src/i18n.ts
@@ -1,0 +1,14 @@
+import { notFound } from 'next/navigation';
+import { getRequestConfig } from 'next-intl/server';
+
+// Can be imported from a shared config
+const locales = ['en', 'es', 'pt'];
+
+export default getRequestConfig(async ({ locale }) => {
+  // Validate that the incoming `locale` parameter is valid
+  if (!locales.includes(locale as any)) notFound();
+
+  return {
+    messages: (await import(`./locales/${locale}.json`)).default,
+  };
+});
