+++ src/middleware.ts
@@ -1,0 +1,14 @@
+import createMiddleware from 'next-intl/middleware';
+
+export default createMiddleware({
+  // A list of all locales that are supported
+  locales: ['en', 'es', 'pt'],
+  localeDetection: false,
+  // Used when no locale matches
+  defaultLocale: 'en',
+});
+
+export const config = {
+  // Match only internationalized pathnames
+  matcher: ['/', '/(pt|es|en)/:path*'],
+};
