+++ src/navigation.ts
@@ -1,0 +1,6 @@
+import { createSharedPathnamesNavigation } from 'next-intl/navigation';
+import { locales } from './constants/locales';
+
+export const { Link, redirect, usePathname, useRouter } = createSharedPathnamesNavigation({
+  locales,
+});
